requirements.txt
langchain
langchain-community
langchain-openai
azure-core>=1.30.0
azure-search-documents>=11.6.0
azure-identity>=1.16.0

msal
requests
python-dotenv
pypdf
python-docx
python-pptx
openpyxl
fastapi
uvicorn
tiktoken
langchain>=0.2.10
langchain-core>=0.2.10
langchain-community>=0.2.10
langchain-openai>=0.2.0
langchain-text-splitters>=0.2.0


myapplication